Documentation for ITensorsOpenSystems.GlobalFunctionApproximation

ITensorsOpenSystems.GlobalFunctionApproximation.augustbanuls — Function

augustbanuls(funcs,A,U0,maxiter;kwargs...)

augustbanuls(
func::Function,
A::ITensors.MPO,
U0::ITensors.MPO,
maxiter::Int;
kwargs...,

) Evaluates functions of the form Tr[f(A)], where A is a Hermian operator given as an MPO and f is an arbirary function, using the algorithm in August,Banuls (2018).

Accepts either a single function or a list of functions [f,g,h...] to evaluate Tr[f(A)],Tr[g(A)],Tr[h(A)]... simultaniously, as this is substantially more efficent that running the algorithm separately for each function.

#Arguments funcs: The function or list of functions to be evaluated on the target matrix A: The argument Hermitian matrix on which the functions are to act U0: A unitary matrix, given as an MPO with the same set of indices as m which acts as an inital value of the iteration maxiter: The maximum number of loops of the algorithm, after which it will terminate and return the result, even if the desired convergence has not been achieved #Key word arguments tol: The maximum desired error for the algorithm to terminate cuttoff, maxdim, mindim: Determine the truncation conditions for the MPO operations in the usual way, the same as for built in ITensors functions.