QuantumTrajectories

ITensorsOpenSystems.QuantumTrajectories.trajectory — Method

trajectory(evolve::<:Function, jumpops, ψ0::MPS,jumptimes)::Vector{MPS}

Calculates the evolution of the states ψ0 along a single jump trajectory. The function evolve(ψ,t1,t2) time evolves the state ψ from time t1 to time t2 with no jumps taking place. The jump operators, jumpops, are given as an iterable of opjects that can be applyed to an MPS with ITensors.apply, i.e. MPOs, ITensors or vectors of ITensors. The state is evolved from jumptimes[1] to jumptimes[end], checking if a jump should occure at each jumptime.

Returns the state at each point in time along the trajectory in sampletimes

##Example

import Dates:second,now
using ITensors
using ITensorTDVP

N = 5
J = 1.0
γ = 0.5

indices = siteinds("S=1/2",N)

Hop = OpSum()

# Hermitian part of the Hamiltonian
for i = 1:(N-1)
    Hop += J, "Sz", i, "Sz", i+1
end

# Non-Hermitian part of the Hamiltonian
Hop += -im*γ/2, "Sp", 1, "Sm", 1
Hop += -im*γ/2, "Sm", N, "Sp", N

H = MPO(Hop,indices)

# Define the jump operators corresponding
# the non Hermitian part of H
jumpops = [γ * op("Sm",1), γ * op("Sp",N)]

ψ0 = randomMPS(indices)

# We want to evolve the systems for 0 to tfinal, with potential jumps every tstep
tfinal = 5
tstep = 0.5

#Seed the random number generator
seed!(second(now()))

# Final calculate the trajectory
trajectory_ψ = trajectory(jumpops, ψ0, 0:tstep,tfinal) do ψ, t1, t2
    # Between jumps, evolve the system under H by TDVP
    return tdvp(H,ψ,(t2-t1);cutoff=1e-7,normalize=false)
end

see also: quantumtrajectories