Vectorization

ITensorsOpenSystems.Vectorization.anticommutatorapply — Method

anticommutatorapply(q)

Constructs the vectorized operator representation of the operator
anticommutator {q,⋅} as an autoMPO object

see also: `commutatorapply`, `leftrightapply`

ITensorsOpenSystems.Vectorization.commutatorapply — Method

commutatorapply(q)

Constructs the vectorized operator representation of the operator
commutator [q,⋅] as an autoMPO object

see also: `anticommutatorapply`, `leftrightapply`

ITensorsOpenSystems.Vectorization.dissipator — Method

dissipator(opsum::Vector, fatsys)

# Arguments
- `opsum::Vector`: Vector of OpSums
- `fatsys`: Vectorized system

# Examples

...

ITensorsOpenSystems.Vectorization.fatsiteinds — Method

fatsiteinds(tag::String, N::Integer; kwargs...)

# Arguments
- `tag::String`: the name of the base space to vectorize
- `N::Integer`: the number of site indices to create
- `operators::Vector{String}`: (Optional) A vector of operators

# Examples
```
julia> fatsys = fatsiteinds("S=1/2", 3)
3-element Vector{ITensors.Index{Int64}}:
(dim=4|id=770|"S=1/2Vec,Site,n=1")
(dim=4|id=704|"S=1/2Vec,Site,n=2")
(dim=4|id=428|"S=1/2Vec,Site,n=3")

julia> operators = ["Z", "X", "Y", ("Sp", "Sm")]
4-element Vector{Any}:
"Z"
"X"
"Y"
("Sp", "Sm")

julia> fatsys = fatsiteinds("S=1/2", 3, operators)
3-element Vector{ITensors.Index{Int64}}:
(dim=4|id=625|"S=1/2Vec,Site,n=1")
(dim=4|id=31|"S=1/2Vec,Site,n=2")
(dim=4|id=697|"S=1/2Vec,Site,n=3")
```

Create an array of N physical site indices of type tag within an vectorized space. Keyword arguments can be used to specify quantum number conservation, see the space function corresponding to the site type tag for supported keyword arguments. These keyword arguments are passed through to the ITensors.siteinds function.

ITensorsOpenSystems.Vectorization.operatorconj — Method

operatorconj(rho::VectorizedDensityMatrix)::VectorizedDensityMatrix

Performs the hermitian conjugate on a vectorized density matrix

ITensorsOpenSystems.Vectorization.unvectorize_density_matrix! — Method

unvectorize_density_matrix!(m::ITensors.MPS,unvectorized_inds::Vector{<:ITensors.Index})::ITensors.MPO

Undoes the vectorization of a VectorizedDensityMatrix, in place, resulting in an MPO. The indices of the resulting MPO are given by unvectorized_inds

ITensorsOpenSystems.Vectorization.vectorize_density_matrix! — Method

vectorize_density_matrix!(o::ITensors.MPO,vectorizedinds::Vector{<:ITensors.Index})::VectorizedDensityMatrix

Treating the MPO as a density matrix, performs fat vectorization in place, resulting in a VectorizedDensityMatrix with indices given by vectorizedinds

see also: VectorizedDensityMatrix

ITensorsOpenSystems.Vectorization.vectorizedexpect! — Method

vectorizedexpect!(A::ITensors.MPO,ρ::ITensors.MPS)

Evaluates the expecation value Tr(Aρ), where A is an MPO and ρ is the vectorized density matrixrepresented as an MPS.

Note that A is not a vectorized operator here and should act on the unvectorized space.

This function will modify the operator A, but leave the state ρ unchanged

ITensorsOpenSystems.Vectorization.vectorizedexpect — Method

vectorizedexpect(A::ITensors.MPO,ρ::ITensors.MPS)

Evaluates the expecation value Tr(Aρ), where A is an MPO and ρ is the vectorized density matrixrepresented as an MPS.

Note that A is not a vectorized operator here and should act on the unvectorized space.

ITensorsOpenSystems.Vectorization.vectorizedexpect — Method

vectorizedexpect(operator::Union{String,Matrix},vectorisedstate::ITensors.MPS)# TODO makes sites into a keyword to match ITensors.expect

...

ITensorsOpenSystems.Vectorization.vectorizedexpect — Method

vectorizedexpect(operator::Union{ITensors.Op,ITensors.Prod{ITensors.Op},ITensors.Sum{ITensors.Scaled}},vectorisedstate::ITensors.MPS)

...

ITensorsOpenSystems.Vectorization.MPSClones.@cloneMPS — Macro

Creates a struct, which is a subclass of ITensors.AbstractMPS, which implements the interface that, as far as possible, implements the interface that ITensor assumes for AbstractMPSs