Documentation for ITensorsOpenSystems._InternalUtils

ITensorsOpenSystems._InternalUtils.autoMPO_multiply — Function

autoMPOmultiply(a::Scaled,b::Prod)::Scaled autoMPOmultiply(a::Sum,b::Sum) autoMPOmultiply(a::Sum,b) autoMPOmultiply(a,b::Sum) autoMPO_multiply(a, b)

Multiplies a pair of autoMPO objects in such a way that the result is an "expected form" That is it will always produce a sum of products, rather than a product of sums and it will correctly multiply a Scaled by a Prod

ITensorsOpenSystems._InternalUtils.partition_kwargs — Function

function partition_kwargs(kwargs,sigils::String...)::Dict{String,<:Dict}

Separates a named tuple into subsets based on suffixes of the keywords
returns a dictionary taking the suffixes as the keys and with a dictionary
of keywords (with the suffix removed) and associated values

ITensorsOpenSystems._InternalUtils.partition_nosigil — Constant

const partition_nosigil = "No Sigil" unsure exactly what this

ITensorsOpenSystems._InternalUtils.partitionsets — Function

partitionsets(s1::Set, s2::Set)::Tuple{Set,Set,Set}

Given two sets, setarates the elements into those that are only in the first, those that are in both and those that are only in the second